If given a local file path to a transcript, this will parse the transcript PDF and extract CS and math classes.
CS and math classes are put into a transcript array and have a 'name' and 'grade'. Name is something like COMPSCI 220 and grades are letter grades with + or - if applicable. To test on a different transcript, change the file path. Next task is to set this up using Multer so that it connects to the server who will provide it with a path to the user's transcript.